Major overhaul. Renamed PlayerGameInfo to PlayerComponent, and redesigned scope so that objects such as camera and input managers are local-only components that are added via Player.AddPlayerComponent<T>()
Implemented the notion of a GameSession. This solves the problem of having the PoolGameDirector wait for players to join
and synchronize before starting the game. From hereon, ISessionHandler.OnPlayerJoined will be called whenever a player
joins and successfully acknowledges the network game. As a result, upper layers need not come up with ways to deal with
network synchronization. How and when the GameSession object is created is still WIP.

Implemented inputs for handling the rack, and streamlined input handling in general.

Surrounding changes in the PoolGameDirector and TurnController to support the new session flow.

Added helper functions and objects in PoolTable to better help us pinpoint where the table surface is, and where other key locations
are (i.e cue ball placement).

Next steps: Move ownership of Cue back to server, and follow same model as the rack. Greatly simplifies state synchronization,
and is more secure. Inputs to the Cue do not work at the moment.