Added the basis of resource management system via ResourceMgr. The ResourceMgr addresses this simple yet devastating problem: - Given an ID for an asset, what is the corresponding resource to load?
Unity only allows to load resources at runtime only if they are placed in a Resources folder - runtime resource loading
is essential for replicating changes in an asset for multiplayer. The folder requirements also restricts us to a predefined
folder structure. This solution uses a custom tool (BuildResourceIndex), which iterates over all assets in the Resources folder
and writes an index file that maps their name hash to the full path. During runtime, we hash the name of an asset, replicate the ID,
and use the index to retrieve the actual resource path we want to load using Resources.Load()... such an interface is abstracted
behind the ResourceMgr singleton.